requirements.txt
python-dotenv==1.0.0
openai==0.28
tweepy==4.14.0
python-dateutil==2.8.2 